Setting Maximum threshold for obtaining plan
For the planning module we are considering that total possible ways to
spend should always be less than 1000. This protects the network from
any DDoS Attack if in case you receive a very large descriptor enough
that it has 1000 possible spend paths.

+++ src/descriptor/tr.rs
@@ -168,7 +168,7 @@ impl TapTree<DescriptorPublicKey> {
     }
 
     /// Get total possible assets for TapTree
-    pub fn count_assets(&self) -> u64 {
+    pub fn count_assets(&self) -> u32 {
         match self {
             TapTree::Tree { left, right, height: _ } => {
                 let a = left.count_assets();
